--- app/src/main/java/com/leondzn/starter/di/modules/ViewModelModule.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2019 Leonard Dizon.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.leondzn.starter.di.modules
-
-import dagger.Module
-
-/**
- * Declare all ViewModels here
- */
-@Module
-abstract class ViewModelModule {
-  //@Binds
-  //@IntoMap
-  //@ViewModelKey(SampleViewModel::class)
-  //abstract fun bindViewModel(viewModel: SampleViewModel): ViewModel
-}
