+++ app/src/main/java/com/snekbyte/starter/ViewModelFactory.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2019 Leonard Dizon.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.snekbyte.starter
+
+import android.arch.lifecycle.ViewModel
+import android.arch.lifecycle.ViewModelProvider
+import javax.inject.Inject
+import javax.inject.Provider
+
+@Suppress("UNCHECKED_CAST")
+class ViewModelFactory
+@Inject constructor(private val viewModelsMap: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>) :
+    ViewModelProvider.Factory {
+
+  override fun <T : ViewModel> create(modelClass: Class<T>): T {
+    val creator = viewModelsMap[modelClass]
+        ?: viewModelsMap.asIterable().firstOrNull { modelClass.isAssignableFrom(it.key) }
+            ?.value ?: throw IllegalArgumentException("Unknown model class $modelClass")
+
+    return creator.get() as T
+  }
+}
