+++ data/src/main/AndroidManifest.xml
@@ -15,5 +15,5 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.leondzn.starter.data">
+    package="com.snekbyte.starter.data">
 </manifest>
